move uis to appropriate folders
