+++ qelem/lib/presentation/pages/home_screen/home_screen.dart
@@ -1,9 +1,8 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
+import '../../../domain/models/question_model.dart';
+import '../question/widgets/question_card.dart';
 
-import '../domain/models/question_model.dart';
-import '../widgets/question_card.dart';
-
-Widget myQuestionsScreen() {
+Widget homeScreen() {
   return ListView.builder(
     itemCount: 20,
     itemBuilder: (context, position) {
