--- qelem/lib/screens/screens.dart
@@ -1,9 +1,0 @@
-export 'package:qelem/screens/main_screen.dart';
-export './change_password_screen.dart';
-export './edit_profile_screen.dart';
-export './register_login_screen.dart';
-export './registration_screen.dart';
-export './login_screen.dart';
-export './post_question_screen.dart';
-export './edit_question_screen.dart';
-export './question_detail_screen.dart';
