+++ qelem/test/widget_test.dart
@@ -6,7 +6,7 @@
 // tree, read text, and verify that the values of widget properties are correct.
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:qelem/main.dart';
+import 'package:qelem/presentation/app_widget.dart';
 
 void main() {
   testWidgets('Home widget test', (WidgetTester tester) async {
